+++ src/ts/tools/events/index.ts
@@ -1,3 +1,4 @@
 export { default as insufficientLight } from './insufficientLight';
 export { default as oldEventValue } from './oldEventValue';
+export { default as bombLighting } from './unlitBomb';
 export { default as testEvent } from './testEvent';